pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyantispam"
version = "0.1.0"
description = "Intelligent email spam detection and filtering system"
authors = [{name = "Pierre", email = "pierre@example.com"}]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "python-dotenv>=1.0.0",
    "email-validator>=2.0.0",
    "scikit-learn>=1.3.0",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "requests>=2.31.0",
]

[project.scripts]
pyantispam = "pyantispam.cli:main"

[project.urls]
Homepage = "https://github.com/pierre/pyantispam"
Repository = "https://github.com/pierre/pyantispam"